requirements.txt
Scrapy>=2.11.0  # Ensure you are using a compatible version of Scrapy
psycopg2-binary
redis==3.5.3
pymongo==3.11.4  # Optional for MongoDB
twisted==21.2.0  # adjust as necessary
pyOpenSSL>=21.0.0